data_center/pdc_backend/requirements.txt
# Core
Django==5.0.4
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
django-filter==24.1

# Database
psycopg2-binary==2.9.9

# Cache / Task Queue
redis==5.0.3
celery==5.3.6
django-celery-beat==2.6.0
django-celery-results==2.5.1

# Storage (Cloudflare R2 via S3-compatible)
boto3==1.34.74
django-storages==1.14.2

# API Docs
drf-spectacular==0.27.2

# AI
google-generativeai==0.5.4

# PDF Generation
reportlab==4.1.0
weasyprint==61.2

# Image Processing
Pillow==10.3.0
# rembg==2.0.57  # Optional: requires separate installation in production.
#                # In production, the Remove.bg API is used as a fallback
#                # if rembg is not installed. To use rembg locally, install
#                # it manually: pip install rembg==2.0.57

# Environment
python-decouple==3.8
dj-database-url==2.1.0

# Utilities
celery[redis]==5.3.6
django-extensions==3.2.3
gunicorn==21.2.0
whitenoise==6.6.0
openpyxl
